Handle deletion during pause

+++ controllers/virtualmachine/virtualmachine_controller.go
@@ -32,6 +32,7 @@ import (
 	"github.com/vmware-tanzu/vm-operator/pkg/patch"
 	"github.com/vmware-tanzu/vm-operator/pkg/prober"
 	"github.com/vmware-tanzu/vm-operator/pkg/providers"
+	vspherevm "github.com/vmware-tanzu/vm-operator/pkg/providers/vsphere/virtualmachine"
 	"github.com/vmware-tanzu/vm-operator/pkg/record"
 	kubeutil "github.com/vmware-tanzu/vm-operator/pkg/util/kube"
 )
@@ -265,12 +266,24 @@ func requeueDelay(ctx *pkgctx.VirtualMachineContext) time.Duration {
 func (r *Reconciler) ReconcileDelete(ctx *pkgctx.VirtualMachineContext) (reterr error) {
 	ctx.Logger.Info("Reconciling VirtualMachine Deletion")
 
+	// If the VM reconciliation has been paused by the developer,
+	// skip deletion and return.
+	if _, exists := ctx.VM.Annotations[vmopv1.PauseAnnotation]; exists {
+		ctx.Logger.Info("VirtualMachine is not deleted because devops paused this VM")
+		return nil
+	}
+
 	if controllerutil.ContainsFinalizer(ctx.VM, finalizerName) {
 		defer func() {
 			r.Recorder.EmitEvent(ctx.VM, "Delete", reterr, false)
 		}()
 
 		if err := r.VMProvider.DeleteVirtualMachine(ctx, ctx.VM); err != nil {
+			// If VM can not be deleted due to reconciliation being paused, ignore that.
+			if errors.Is(err, vspherevm.ErrorVMPausedByAdmin()) {
+				ctx.Logger.Info("VM could not be deleted since it contains the pause reconcile ExtraConfig key")
+				return nil
+			}
 			ctx.Logger.Error(err, "Failed to delete VirtualMachine")
 			return err
 		}
